product_search/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
google-generativeai
pandas
numpy

